requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
coverage==7.9.2
Django==5.2.4
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
django-environ==0.12.0
requests==2.32.4
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
